# gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AdminLoginErrorResponse:
    properties:
      error:
        example: <error_message>
        type: string
      message:
        example: Failed to Login Admin!
        type: string
    type: object
  models.AdminLoginSuccessResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjdhZGNhMGI0ZTg2NGNmZmIwMDIyOTkiLCJleHBpcmVkX2F0IjoiMjAyMS0wNy0wMVQwMDowMDowMC4wMDBaIiwiaWF0IjoxNjI0MzQwNjAwfQ.7
        type: string
      admin:
        $ref: '#/definitions/models.Employee'
      message:
        example: Admin Logged In Successfully!
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjdhZGNhMGI0ZTg2NGNmZmIwMDIyOTkiLCJleHBpcmVkX2F0IjoiMjAyMS0wNy0wMVQwMDowMDowMC4wMDBaIiwiaWF0IjoxNjI0MzQwNjAwfQ.7
        type: string
    type: object
  models.BrandAccessDTO:
    properties:
      brand_id:
        type: string
      roles:
        items:
          type: string
        type: array
    required:
    - brand_id
    - roles
    type: object
...
  models.Employee:
    properties:
      access:
        items:
          $ref: '#/definitions/models.BrandAccessDTO'
        type: array
      created_at:
        type: string
      email:
        example: johndoe@example.com
        type: string
      id:
        type: string
      incentive:
        type: integer
      is_active:
        type: boolean
      name:
        example: johndoe
        type: string
      phone:
        example: 9876543210
        maximum: 10
        minimum: 10
        type: integer
      role:
        example: employee
        type: string
      updated_at:
        type: string
      username:
        example: johndoe
        maxLength: 20
        minLength: 3
        type: string
    required:
    - email
    - name
    - phone
    - username
    type: object
...
  models.UserLoginPayload:
    properties:
      identifier:
        example: johndoe/johndoe@example.com
        type: string
      password:
        example: strong@password
        type: string
    required:
    - identifier
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: rohit.vilas.ingole@gmail.com
    name: Rohit Vilas Ingole
  description: This is the API for Gopher Social, a social media platform for Gophers.
  license:
    name: MIT License
    url: https://github.com/DataRohit/Order-Management-System/blob/master/license
  title: Gopher Social API
  version: "1.0"
paths:
...
  /auth/admin/login:
    post:
      consumes:
      - application/json
      description: Login an Admin to the Platform
      parameters:
      - description: Admin Login Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.UserLoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in admin
          schema:
            $ref: '#/definitions/models.AdminLoginSuccessResponse'
        "400":
          description: Failed to login the admin
          schema:
            $ref: '#/definitions/models.AdminLoginErrorResponse'
        "500":
          description: Failed to login the admin
          schema:
            $ref: '#/definitions/models.AdminLoginErrorResponse'
      summary: Login Admin
      tags:
      - Auth
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"